third_party/blink/web_tests/fast/js/function-prototype-expected.txt

Tests Bugzilla Bug 9815: JavaScript TypeError loading Dean Edwards' JS compressor/obfuscator

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS actual is 'function() { }'
PASS successfullyParsed is true

TEST COMPLETE
